backend/requirements.txt
# =================================================================
# CORE FRAMEWORK & WEB SERVER
# =================================================================
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
gunicorn==21.2.0

# =================================================================
# CONFIGURATION & ENVIRONMENT MANAGEMENT
# =================================================================
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
dynaconf==3.2.4

# =================================================================
# DATABASE & ORM
# =================================================================
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter (production)
alembic==1.13.1
SQLAlchemy-Utils==0.41.1

# =================================================================
# REAL-TIME COMMUNICATION
# =================================================================
python-socketio==5.11.0
eventlet==0.33.3
redis==5.0.1
celery==5.3.4

# =================================================================
# SECURITY & AUTHENTICATION
# =================================================================
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
cryptography==45.0.4
passlib==1.7.4
argon2-cffi==23.1.0

# =================================================================
# API & VALIDATION
# =================================================================
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
webargs==8.3.0
apispec==6.3.0
flask-apispec==0.11.4

# =================================================================
# MONITORING & OBSERVABILITY
# =================================================================
prometheus-flask-exporter==0.23.0
structlog==23.2.0
elastic-apm==6.19.0
sentry-sdk[flask]==1.38.0

# =================================================================
# CACHING & PERFORMANCE
# =================================================================
Flask-Caching==2.1.0
redis==5.0.1
python-memcached

# =================================================================
# GEOSPATIAL PROCESSING
# =================================================================
Shapely==2.0.2
geojson==3.1.0
geoalchemy2==0.14.2
pyproj==3.6.1

# =================================================================
# AI & MACHINE LEARNING
# =================================================================
numpy==1.26.4 
scikit-learn==1.3.0
scipy==1.11.1
joblib==1.3.2
mlflow==2.8.1

# =================================================================
# DATA ANALYTICS & VISUALIZATION
# =================================================================
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# =================================================================
# EXTERNAL SERVICES & APIS
# =================================================================
requests==2.31.0
requests-cache==1.1.0
boto3==1.34.0  # AWS SDK
google-cloud-storage==2.10.0  # Google Cloud
azure-storage-blob==12.19.0  # Azure Storage

# =================================================================
# UTILITIES & DATA PROCESSING
# =================================================================
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
python-slugify==8.0.1
email-validator==2.1.0

# =================================================================
# TESTING & QUALITY ASSURANCE
# =================================================================
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
factory-boy==3.3.0
flask-testing==0.8.1
coverage==7.3.2

# =================================================================
# DEVELOPMENT TOOLS
# =================================================================
flask-shell-ipython==1.4.0
ipython==8.17.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# =================================================================
# TASK QUEUE & BACKGROUND JOBS
# =================================================================
celery==5.3.4
flower==2.0.1  # Celery monitoring

# =================================================================
# HEALTH CHECKS & DIAGNOSTICS
# =================================================================
py-healthcheck==1.10.1

# =================================================================
# DEPLOYMENT & CONTAINERIZATION
# =================================================================
docker==6.1.3
kubernetes==28.1.0

# =================================================================
# ENVIRONMENT-SPECIFIC DEPENDENCIES
# =================================================================
# Production
supervisor==4.2.5  # Process management

# Development
flask-debugtoolbar==0.13.1
werkzeug==3.0.1